Update docs after further discussion and exploration
Tom pointed out why version.pm was dying silently some of the time and
made it appear that VERSION wasn't getting called, so I've updated the
docs to reflect that.

We *could* rewrite this to check whether or not our version is parsable
before overriding UNIVERSAL::VERSION but somehow that seems worse than
the current code.  We know we use version numbers that are incompatible
with perl version numbers, but we aren't going to stop naming things rc1
at this point.